requirements.txt
Django >3.2,<3.3
python-decouple
Pillow
djangorestframework
markdown
django-filter
drf-yasg[validation]
djangorestframework-simplejwt
django-mail-templated
djoser
black
flake8